% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix.gen}
\alias{ld_matrix.gen}
\title{If structured = TRUE: Nei & Li (1973) DOI: 10.1093/genetics/75.1.213}
\usage{
\method{ld_matrix}{gen}(
  x,
  measure = c("r2", "r2hap", "D'", "D'hap", "Chisq", "Chisqdf"),
  structured = TRUE,
  ...
)
}
\description{
If structured = TRUE: Nei & Li (1973) DOI: 10.1093/genetics/75.1.213
}
